Sap finish

+++ app/Http/Controllers/UploadFileController.php
@@ -9,6 +9,9 @@ use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\File;
 use App\Http\Controllers\Response;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Hash;
+use SimpleSoftwareIO\QrCode\Facades\QrCode;
+use Barryvdh\DomPDF\Facade as PDF;
 
 class UploadFileController extends Controller
 {
@@ -18,24 +21,36 @@ class UploadFileController extends Controller
     }
     public function fileUpload(Request $request)
     {
+        dd(url($path = 'app/'.'public/' . 'uploads' . '/' . 'abc'));
+        // $randomString = Str::random(7);
+        // $path = 'app/'.'public/' . 'uploads' . '/' . $randomString;
+        // $qrCodeData = base64_encode(QrCode::format('png')->size(300)->generate('abc'));
+        // if (!File::exists(storage_path($path))) {
+        //     File::makeDirectory($path, 0755, true, true);
+        // }
+        // file_put_contents(storage_path($path), $qrCodeData);        
+        //  return redirect()->route('show-form-uploadFile')->with('success', 'Multiple File has been uploaded Successfully')
+        // ->with('image', $qrCodeData);
 
-        // $validatedData = $request->validate([
-        //     'files' => 'required',
-        //     'files.*' => 'mimes:csv,txt,xlx,xls,pdf'
-        // ]);
-        //lấy đường dẫn
-        $path = 'public/' . 'uploads' . '/' . $request->session()->get('id') . '/' . $request->name;
-        if (!File::exists(storage_path($path))) {
-            File::makeDirectory($path, 0755, true, true);
-        }
 
+        $validatedData = $request->validate([
+            'files' => 'required',
+            'files.*' => 'mimes:csv,txt,xlx,xls,pdf,doc,docx'
+        ]);
         if (!$request->session()->has('id')) {
             //     //tạo ra QR code rồi lưu vào dự án thông qua httplocalhost
             //     //địa chỉ storage
-            return redirect()->route('show-form-login');
-        }
-        //  $check = url('storage/'.'app/'.$path);
-        //  dd($check);
+            $randomString = Str::random(7);
+            $path = 'public/' . 'uploads' . '/' . $randomString.'/'.$request->name;
+            if (!File::exists(storage_path($path))) {
+                File::makeDirectory($path, 0755, true, true);
+            }
+        }else{
+            $path = 'public/' . 'uploads' . '/' . $request->session()->get('id') . '/' . $request->name;
+            if (!File::exists(storage_path($path))) {
+                File::makeDirectory($path, 0755, true, true);
+            }  
+        }              
         $fileBag = $request->file('files');
         if ($fileBag) {
             // Lấy danh sách các file
@@ -49,28 +64,59 @@ class UploadFileController extends Controller
                 print_r($path_pdf);
                 $allFilePaths .= $path_pdf . '|';
             }
-            $file_db = new Files();
-            $file_db->name = $request->name;
-            $file_db->filepath = $allFilePaths;
-            $file_db->userid = $request->session()->get('id');
-            $file_db->save();
-            $fileId = $file_db->id;
+            if($request->session()->has('id')){
+                $file_db = new Files();
+                $file_db->name = $request->name;
+                $file_db->filepath = $allFilePaths;
+                $file_db->userid = $request->session()->get('id');
+                $file_db->save();
+                $fileId = $file_db->id;
+    
+                // Sử dụng ID để tạo link QRCode
+                $file_db->qrcode = url('storagefile/' . $fileId . '/edit');
+                $file_db->save();
+            }
+            //download-QRCODE
+            //  $path_qr =  QrCode::size(100)->generate($file->qrcode); 
 
-            // Sử dụng ID để tạo link QRCode
-            $file_db->qrcode = url('storagefile/' . $fileId . '/edit');
-            $file_db->save();
+            // if (File::exists($path)) {
+            //     return response()->download($path);
+            // } else {
+            //     return "File not found";
+            // }
+            $qrCodeData = QrCode::format('png')->size(300)->generate('abc');
+          
+        return redirect()->route('show-form-uploadFile')->with('qrcode', $qrCodeData);
+         
         }
         return redirect()->route('show-form-uploadFile')->with('success', 'Multiple File has been uploaded Successfully');
     }
     public function allfile()
     {
-        $file = Files::all();
+        if (!session('id')) {
+            //     //tạo ra QR code rồi lưu vào dự án thông qua httplocalhost
+            //     //địa chỉ storage
+            return redirect()->route('show-form-login');
+        }
+        $file = DB::table('files')->where('userid', session('id'))->get();
 
         return view('show-file.index', compact('file'));
     }
     public function show_edit($id)
     {
-        $file = DB::table('files')->where('id', $id)->first();
+        $files = Files::find($id);
+        if($files->userid !== session('id')){
+            abort(404);
+        }
+        if (!session('id')) {
+            //     //tạo ra QR code rồi lưu vào dự án thông qua httplocalhost
+            //     //địa chỉ storage
+            return redirect()->route('show-form-login');
+        }
+        $file = DB::table('files')
+            ->where('id', '$id')
+            ->where('userid', session('id'))
+            ->get();
         $files = File::files(storage_path('app/public/uploads/' . session('id') . '/' . $file->name));
         return view('show-file.edit', compact('file', 'files'));
     }
@@ -78,15 +124,13 @@ class UploadFileController extends Controller
     {
     }
 
-
-    public function download(Request $request)
-    {
-        // $path = storage_path('app/public/uploads/'.session()->get(''));
-        $request->session()->put('path', $request->path);
-        return redirect()->route('ondownload');
-    }
     public function ondownload(Request $request)
     {
+        if (!session('id')) {
+            //     //tạo ra QR code rồi lưu vào dự án thông qua httplocalhost
+            //     //địa chỉ storage
+            return redirect()->route('show-form-login');
+        }
         $path2 = storage_path('app\public\uploads/' . session()->get('id') . '/' . $request->name . '/' . $request->path);
         if (File::exists($path2)) {
             return response()->download($path2);
@@ -96,6 +140,11 @@ class UploadFileController extends Controller
     }
     public function ondelete(Request $request)
     {
+        if (!session('id')) {
+            //     //tạo ra QR code rồi lưu vào dự án thông qua httplocalhost
+            //     //địa chỉ storage
+            return redirect()->route('show-form-login');
+        }
         $path2 = storage_path('app\public\uploads/' . session()->get('id') . '/' . $request->name . '/' . $request->path);
         if (File::exists($path2)) {
             File::delete($path2);
@@ -106,6 +155,11 @@ class UploadFileController extends Controller
     }
     public function deleteall(Request $request, $id)
     {
+        if (!session('id')) {
+            //     //tạo ra QR code rồi lưu vào dự án thông qua httplocalhost
+            //     //địa chỉ storage
+            return redirect()->route('show-form-login');
+        }
         $exsitsFile = DB::table('files')->where('id', $id)->first();
         if (!$exsitsFile) {
             return null;
@@ -119,6 +173,16 @@ class UploadFileController extends Controller
     }
     public function doedit(Request $request)
     {
+        $files = Files::find($request->id);
+        if($files->userid !== session('id')){
+            abort(404);
+        }
+        if (!session('id')) {
+            //     //tạo ra QR code rồi lưu vào dự án thông qua httplocalhost
+            //     //địa chỉ storage
+            return redirect()->route('show-form-login');
+        }
+       
         $nameOld = $request->get('nameOld');
         $nameNew = $request->get('nameNew');
         $Oldpath = storage_path('app\public\uploads/' . session()->get('id') . '/' . $nameOld);
@@ -147,5 +211,11 @@ class UploadFileController extends Controller
             }
         }
         return redirect()->route('storagefile');
+    }
+    public function show_edit_anonymous(Request $request){       
+        $parts = explode('/', $request->getPathInfo());
+        $info = $parts[1];
+        $files = File::files(storage_path('app/public/uploads/' .  $info));
+        dd($files);  
     }
 }
